+++ src/it/it-withOption-exactOutput-target-overwrite/src/main/java/name/valery1707/kaitai/it/Ico.java
@@ -1,0 +1,8 @@
+package name.valery1707.kaitai.it;
+
+/**
+ * Stub for test overwriting by generated classes.
+ */
+public class Ico {
+
+}